% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{add_cols}
\alias{add_cols}
\title{Add columns if columns don't exist}
\usage{
add_cols(dset, ds_part, target_cols)
}
\arguments{
\item{dset}{A data frame. The dataset.}

\item{ds_part}{A character scalar. Prefix or postfix.}

\item{target_cols}{A character vector with necessary columns.}
}
\value{
A data frame. The dataset.
}
\description{
Add columns if columns don't exist
}
