% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{calc_diff}
\alias{calc_diff}
\title{Function for calculating the difference between two dates}
\usage{
calc_diff(st_inter, dt_item)
}
\arguments{
\item{st_inter}{An interval. An object of interval.}

\item{dt_item}{A date item. An object of date.}
}
\value{
An integer scalar. Differences between the two dates.
}
\description{
Function for calculating the difference between two dates
}
