% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract.R
\name{check}
\alias{check}
\title{Check the dataset}
\usage{
check(obj, dataset)
}
\arguments{
\item{obj}{An object for check.}

\item{dataset}{A dataset, a type is a data frame.}
}
\value{
An object with a check result.
}
\description{
Check the dataset
}
\examples{
id <- c("01", "02", "03")
screen_date_E1 <- c("1991-03-13", "1991-03-07", "1991-03-08")
rand_date_E2 <- c("1991-03-15", "1991-03-11", "1991-03-10")
ph_date_E3 <- c("1991-03-21", "1991-03-16", "1991-03-16")
bio_date_E3 <- c("1991-03-23", "1991-03-16", "1991-03-16")

df <- data.frame(id, screen_date_E1, rand_date_E2, ph_date_E3, bio_date_E3,
  stringsAsFactors = FALSE
)

timeline <- system.file("dates.xlsx", package = "dmtools")
obj_date <- date(timeline, id, dplyr::contains)

obj_date <- check(obj_date, df)
}
