% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{check_sites}
\alias{check_sites}
\title{Check sites}
\usage{
check_sites(objs, dataset, col_site)
}
\arguments{
\item{objs}{A list of objects.}

\item{dataset}{A dataset, a type is a data frame.}

\item{col_site}{A column name of a site in the dataset, without quotes.}
}
\value{
A list of objects with a check result.
}
\description{
Check sites
}
\examples{
SITE <- c("site 01", "site 02")
ID <- c("01", "02")
AGE <- c("19", "20")
SEX <- c("f", "m")
GLUC_V1 <- c("5.5", "4.1")
GLUC_IND_V1 <- c("norm", "no")
AST_V2 <- c("30", "48")
AST_IND_V2 <- c(NA, "norm")

df <- data.frame(
  SITE, ID, AGE, SEX,
  GLUC_V1, GLUC_IND_V1,
  AST_V2, AST_IND_V2,
  stringsAsFactors = FALSE
)

refs_s01 <- system.file("labs_refer_s01.xlsx", package = "dmtools")
refs_s02 <- system.file("labs_refer_s02.xlsx", package = "dmtools")

s01_lab <- lab(refs_s01, ID, AGE, SEX, "norm", "no", site = "site 01")
s02_lab <- lab(refs_s02, ID, AGE, SEX, "norm", "no", site = "site 02")

labs <- list(s01_lab, s02_lab)
labs <- check_sites(labs, df, SITE)
}
