% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab.R
\name{create_norm}
\alias{create_norm}
\title{Estimating laboratory values}
\usage{
create_norm(vals, left_bound, right_bound, ds_norm, normal, abnormal, clsign)
}
\arguments{
\item{vals}{A double vector. The laboratory values.}

\item{left_bound}{A double scalar. The minimum.}

\item{right_bound}{A double scalar. The maximum.}

\item{ds_norm}{An estimate of the laboratory values from the dataset.}

\item{normal}{An option for the normal estimate, for example, "NORMAL".}

\item{abnormal}{An option for the abnormal estimate, for example, "ABNORMAL".}

\item{clsign}{An option for the clinical significant estimate, for example, "CLISIG".}
}
\value{
A vector with the auto estimate.
}
\description{
Estimating laboratory values
}
