% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{create_spec}
\alias{create_spec}
\title{For creating part of the specification}
\usage{
create_spec(df_spec, all_colname, part_spec, is_pst)
}
\arguments{
\item{df_spec}{A dataset, a type is a data frame.}

\item{all_colname}{A character vector with all names in the dataset.}

\item{part_spec}{A character scalar. Prefixes or postfixes.}

\item{is_pst}{A logical scalar, default is TRUE. True is postfix, otherwise, prefix.}
}
\value{
A data frame. Part of the specification.
}
\description{
For creating part of the specification
}
