% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{date}
\alias{date}
\title{Create object date}
\usage{
date(file, id, get_visit, get_date = dplyr::contains, str_date = "DAT")
}
\arguments{
\item{file}{A character scalar. Path to the date's parameters in the excel table.}

\item{id}{A column name of the subject id in the dataset, without quotes.}

\item{get_visit}{A function, which select necessary visit or event e.g. dplyr::start_with, dplyr::contains.}

\item{get_date}{A function, which select dates from necessary visit e.g. dplyr::matches, dplyr::contains, default: dplyr::contains.}

\item{str_date}{A date's pattern in column names, default: "DAT".}
}
\value{
The object date.
}
\description{
Create object date
}
\examples{
obj_date <- date("dates.xlsx", id, dplyr::contains)
obj_date <- date("dates.xlsx", id, dplyr::contains, "uneq")

}
