% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{find_colnames.date}
\alias{find_colnames.date}
\title{Find column names with dates}
\usage{
\method{find_colnames}{date}(obj, dataset, row_file)
}
\arguments{
\item{obj}{An object for validation.}

\item{dataset}{A data frame. Class date.}

\item{row_file}{A data frame. A data frame with analysis parameters.}
}
\value{
A data frame. Visit's dates.
}
\description{
Find column names with dates
}
