% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{find_colnames.default}
\alias{find_colnames.default}
\title{Find column names}
\usage{
\method{find_colnames}{default}(obj, dataset, row_file)
}
\arguments{
\item{obj}{An object for validation.}

\item{dataset}{A dataset, a type is a data frame.}

\item{row_file}{A row of the file.}
}
\value{
A data frame. Result of run_tests.
}
\description{
Find column names
}
