% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract.R
\name{find_colnames}
\alias{find_colnames}
\title{Find column names}
\usage{
find_colnames(obj, dataset, row_file)
}
\arguments{
\item{obj}{An object for check.}

\item{dataset}{A dataset, a type is a data frame.}

\item{row_file}{A row of the file.}
}
\value{
A data frame. Result of run_tests.
}
\description{
Find column names
}
