% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meddra.R
\name{meddra_auth}
\alias{meddra_auth}
\title{Get the token}
\usage{
meddra_auth(target_url, meddra_id, api_key)
}
\arguments{
\item{target_url}{The url for authenticate.}

\item{meddra_id}{The user's meddra id.}

\item{api_key}{The user's api key.}
}
\value{
A string scalar. The user's token.
}
\description{
Get the token
}
\examples{
\dontrun{
meddra_auth(url, id, key)
}
}
