% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meddra.R
\name{meddra_post}
\alias{meddra_post}
\title{Create the post query}
\usage{
meddra_post(target_url, json, token, encoding = "UTF-8")
}
\arguments{
\item{target_url}{The url for a post query.}

\item{json}{A string scalar. The json query.}

\item{token}{The user's token.}

\item{encoding}{The json encoding. The default is utf-8.}
}
\value{
A data frame. The result of query.
}
\description{
Create the post query
}
\examples{
\dontrun{
meddra_post(url, json_body, token)
}
}
