% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{run_tests.date}
\alias{run_tests.date}
\title{Run test}
\usage{
\method{run_tests}{date}(obj, dataset, row_file, date)
}
\arguments{
\item{obj}{An object for validation.}

\item{dataset}{A data frame. Class date.}

\item{row_file}{A data frame. A data frame with analysis parameters.}

\item{date}{A column name with dates.}
}
\value{
A data frame. Result of the date's validation.
}
\description{
Run test
}
