% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract.R
\name{run_tests}
\alias{run_tests}
\title{Run tests}
\usage{
run_tests(obj, dataset, row_file, part)
}
\arguments{
\item{obj}{An object for check.}

\item{dataset}{A data frame.}

\item{row_file}{A data frame. A data frame with parameters.}

\item{part}{A character scalar. Prefixes or postfixes.}
}
\value{
A data frame. The part of final result.
}
\description{
Run tests
}
