% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbc.R
\name{run_tests.wbc}
\alias{run_tests.wbc}
\title{Run tests}
\usage{
\method{run_tests}{wbc}(obj, dataset, row_file, part)
}
\arguments{
\item{obj}{An object. Class wbc.}

\item{dataset}{A data frame.}

\item{row_file}{A data frame. A data frame with parameters.}

\item{part}{A character scalar. Prefixes or postfixes.}
}
\value{
A data frame. The part of the final result.
}
\description{
Run tests
}
