% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short.R
\name{short}
\alias{short}
\title{Create object short}
\usage{
short(file, id, name_to_find, common_cols = c(), extra = NA, is_post = T)
}
\arguments{
\item{file}{A character scalar. Path to the excel table.}

\item{id}{A column name of the subject id in the dataset, without quotes.}

\item{name_to_find}{A character scalar. For search prefixes or postfixes.}

\item{common_cols}{A character vector. A column names in the dataset, which common for all events.}

\item{extra}{A character scalar. For additional information.}

\item{is_post}{A logical scalar, default is TRUE. True is postfix, otherwise, prefix.}
}
\value{
The object short.
}
\description{
Create object short
}
\examples{
obj_short <- short("preg.xlsx", id,"res", c("site", "sex"))
obj_short <- short("labs.xlsx", id,"name_labs", c("site"), "human_name")

}
