% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{test_sites}
\alias{test_sites}
\title{Test sites}
\usage{
test_sites(objs, func)
}
\arguments{
\item{objs}{A list of objects.}

\item{func}{A function e.g. \code{choose_test}, \code{get_result}.}
}
\value{
A data frame. The dataset.
}
\description{
Test sites
}
\examples{
SITE <- c("site 01", "site 02")
ID <- c("01", "02")
AGE <- c("19", "20")
SEX <- c("f", "m")
GLUC_V1 <- c("5.5", "4.1")
GLUC_IND_V1 <- c("norm", "no")
AST_V2 <- c("30", "48")
AST_IND_V2 <- c(NA, "norm")

df <- data.frame(
  SITE, ID, AGE, SEX,
  GLUC_V1, GLUC_IND_V1,
  AST_V2, AST_IND_V2,
  stringsAsFactors = FALSE
)

refs_s01 <- system.file("labs_refer_s01.xlsx", package = "dmtools")
refs_s02 <- system.file("labs_refer_s02.xlsx", package = "dmtools")

s01_lab <- lab(refs_s01, ID, AGE, SEX, "norm", "no", site = "site 01")
s02_lab <- lab(refs_s02, ID, AGE, SEX, "norm", "no", site = "site 02")

labs <- list(s01_lab, s02_lab)
labs <- check_sites(labs, df, SITE)

test_sites(labs, func = function(lab) choose_test(lab, "mis"))
}
