% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{to_dbl}
\alias{to_dbl}
\title{Cast to double type}
\usage{
to_dbl(vals)
}
\arguments{
\item{vals}{A character or double vector.}
}
\value{
A double vector.
}
\description{
Cast to double type
}
