% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short.R
\name{to_long.short}
\alias{to_long.short}
\title{Reshape the dataset to a long view}
\usage{
\method{to_long}{short}(obj, dataset, row_file, part)
}
\arguments{
\item{obj}{An object. Class short.}

\item{dataset}{A data frame.}

\item{row_file}{A data frame. A data frame with parameters.}

\item{part}{A character scalar. Prefixes or postfixes.}
}
\value{
A data frame. The part of the final result.
}
\description{
Reshape the dataset to a long view
}
