% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbc.R
\name{wbc}
\alias{wbc}
\title{Create object wbc}
\usage{
wbc(file, id, is_post = T, site = NA, name_to_find = "relative")
}
\arguments{
\item{file}{A character scalar. Path to the laboratory's reference in the excel table.}

\item{id}{A column name of the subject id in the dataset, without quotes.}

\item{is_post}{A logical scalar, default is TRUE. True is postfix, otherwise, prefix.}

\item{site}{A site number, default: NA.}

\item{name_to_find}{A character scalar. For search prefixes or postfixes, default is "relative".}
}
\value{
The object wbc.
}
\description{
Create object wbc
}
