% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meddra.R
\name{meddra_post}
\alias{meddra_post}
\title{Create the post query}
\usage{
meddra_post(target_url, json, token)
}
\arguments{
\item{target_url}{The url for a post query.}

\item{json}{A string scalar or a list. The json query.}

\item{token}{The user's token.}
}
\value{
A list. The result of query.
}
\description{
Create the post query
}
\examples{
\dontrun{
meddra_post(url, json_body, token)
}
}
